Add files via upload
This code computes the expected value of a fair six sided dice roll. It proves that the expected value from trials get closer to the theoretical value of 3.5, with increasing number of trials (the number of trials range from one to a billion). It's a statistical proof and I love this proof